+++ CHANGES.md
@@ -1,5 +1,20 @@
 ## dev
 
+## 0.7.2 - 2016-03-22
+
+* Hotfix demo fix for 0.7.0 change to `ed.updatePlaceholder` arguments
+
+## 0.7.1 - 2016-03-22
+
+* Update to [ProseMirror 0.5.0](https://discuss.prosemirror.net/t/release-0-5-0/246)
+  * Fixes (re)focus showing empty block menu (#125)
+  * ~~Fixes Android blockers (#29?) (#104?)~~
+
+## 0.7.0 - 2016-03-22
+
+* BREAKING - changed arguments for `ed.updatePlaceholder` to `id, {status, progress, failed}`
+  * Placholder block `metadata.failed` will show in red and have an "×" button to remove
+
 ## 0.6.2 - 2016-03-20
 
 * Hotfix `ed.getContent()` on :iphone: (#105)